requirements.txt
pandas==2.1.4
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
streamlit==1.29.0
plotly==5.17.0
dbt-core==1.7.4
dbt-postgres==1.7.4
python-dotenv==1.0.0
schedule==1.2.0
prefect==2.14.11
prefect>=2.13.0
